move help icon with tooltip to separate component
