+++ src/DashboardUI/src/components/details-page/water-right/WaterRightHeader.tsx
@@ -6,6 +6,7 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { loginRequest } from '../../../authConfig';
 import { isFeatureEnabled } from '../../../config/features';
 import { useState } from 'react';
+import { OverlayTooltip } from '../../OverlayTooltip';
 
 function WaterRightHeader() {
   const navigate = useNavigate();
@@ -13,15 +14,6 @@ function WaterRightHeader() {
   const isAuthenticated = useIsAuthenticated();
   const { instance } = useMsal();
   const [showLoginPromptModal, setShowLoginPromptModal] = useState(false);
-
-  const overlayElement = (props: any) => (
-    <Popover id="consumptive-use-btn-tooltip" {...props}>
-      <Popover.Body>
-        Estimate the amount of water consumed (not returned) based on available data, including water rights, use
-        records, and supply inputs.
-      </Popover.Body>
-    </Popover>
-  );
 
   const consumptiveUseBtnClickHandler = () => {
     if (!isAuthenticated) {
@@ -53,9 +45,10 @@ function WaterRightHeader() {
           <div className="d-flex flex-column align-items-end gap-1">
             <div className="d-flex flex-row gap-3 align-items-center">
               <div>
-                <OverlayTrigger trigger="hover" placement="left" delay={{ show: 0, hide: 0 }} overlay={overlayElement}>
-                  <Icon path={mdiHelpCircleOutline} size="1.5em" />
-                </OverlayTrigger>
+                <OverlayTooltip
+                  text="Estimate the amount of water consumed (not returned) based on available data, including water rights, use
+        records, and supply inputs."
+                />
               </div>
               <div>
                 <Button variant="primary" onClick={consumptiveUseBtnClickHandler}>
